==> Avanced Pandas/02 Preparing Your Dataset.ipynb <==
# Preparing Datasets

import os    # For file paths
import pandas as pd 
from pandas import isnull

# Master.csv
master = pd.read_csv(os.path.join("_datasets", "Master.csv"))
master.head()

master.shape

master.columns

# Braces let us structure the code in a more readable way
(master["playerID"]
     .pipe(isnull)
     .value_counts())
    
# This is an alternative to:
isnull(master["playerID"]).value_counts()

master_orig = master.copy()
master = master.dropna(subset=["playerID"])
master.shape

master.columns

master = master.dropna(subset=["firstNHL", "lastNHL"], how="all")

master = master.loc[master["lastNHL"] >= 1980]
master.shape

master.columns

columns_to_keep = ["playerID", "firstName", "lastName",
                   "pos", "birthYear", "birthMon", "birthDay",
                   "birthCountry", "birthState", "birthCity"]

master[columns_to_keep].head()
master.filter(columns_to_keep).head()

master = master.filter(regex="(playerID|pos|^birth)|(Name$)")
master.columns

def mem_mib(df):
    print("{0:.2f} MiB".format(
        df.memory_usage().sum() / (1024 * 1024)
    ))
    
mem_mib(master)
mem_mib(master_orig)

master["pos"].value_counts()

pd.Categorical(master["pos"])

def make_categorical(df, col_name):
    df.loc[:, col_name] = pd.Categorical(df[col_name]) 

make_categorical(master, "pos")
make_categorical(master, "birthCountry")
make_categorical(master, "birthState")
mem_mib(master)

master.head()

master = master.set_index("playerID")
master.head()

master.to_pickle(os.path.join("_datasets", "master.pickle"))

## Scoring.csv 

scoring = pd.read_csv(os.path.join("_datasets/", "Scoring.csv"))
mem_mib(scoring)
scoring.shape

scoring.columns

def recent_nhl_only(df):
    return df[(df["lgID"] == "NHL") & (df["year"] >= 1980)]

scoring = recent_nhl_only(scoring)
scoring.shape      

scoring.columns

scoring = scoring.filter(regex="^(?!(Post|PP|SH)).*")
scoring.columns

scoring = scoring.iloc[:, [0, 1, 3, 6, 7, 8, 9, 14]]
scoring.columns

make_categorical(scoring, "tmID")
scoring.head()

scoring.reset_index().head()

scoring = scoring.reset_index(drop=True)
# Alternatively:
scoring.reset_index(drop=True, inplace=True)
scoring.head()

scoring.to_pickle(os.path.join("_datasets", "scoring.pickle"))

## Teams.csv 

teams = pd.read_csv(os.path.join("_datasets", "Teams.csv"))
teams.shape

teams.columns

teams = recent_nhl_only(teams)
teams = teams[["year", "tmID", "name"]]
teams.head()

teams.nunique()

make_categorical(teams, "tmID")

teams.to_pickle(os.path.join("_datasets", "teams.pickle"))

## TeamSplits.csv 

team_splits = pd.read_csv(os.path.join("_datasets", "TeamSplits.csv"))
team_splits.shape

team_splits.columns

team_splits = recent_nhl_only(team_splits)

cols_to_drop = team_splits.columns[3:11]
team_splits = team_splits.drop(columns=cols_to_drop)
team_splits.columns
# some_data_frame.drop(rows=row_labels) <- to drop rows

team_splits = team_splits.drop(columns="lgID")

make_categorical(team_splits, "tmID")
team_splits.to_pickle(os.path.join("_datasets", "team_splits.pickle"))
