==> Avanced Pandas/06 Handling Time Series.ipynb <==
import pandas as pd
import os

master = pd.read_pickle(os.path.join("..", "master.pickle"))
team_splits = pd.read_pickle(os.path.join("..", "team_splits_mi.pickle"))

master.head()

master = master.assign(birthDate = pd.to_datetime({
    'year': master.birthYear,
    'month': master.birthMon,
    'day': master.birthDay
}))
master = master.drop(columns=['birthYear', 'birthMon', 'birthDay'])
master.head()

master.birthDate.head(2)

master.birthDate[0]

ts = master.birthDate[0]

print(u"{}, {} {} {}".format(ts.day_name(), ts.day,
                             ts.month_name(),ts.year))


tsz = ts.tz_localize('America/Toronto')
tsz

tsz.tz_convert('Europe/London')

years = master.birthDate.dt.year
years.head()

strings = master.birthDate.dt.strftime("%Y-%d-%Y-%a")
strings.head()

strings = master.birthDate.dt.strftime("%Y-%m-%d")
strings.head()

dates = pd.to_datetime(strings)
dates.head()

strings = strings.str.replace("-", "xx")
strings.head()

try:
    pd.to_datetime(strings)
except Exception as e:
    print(e)

parsed = pd.to_datetime(strings, format="%Yxx%mxx%d")
parsed.head()

strings_error = strings.copy()
strings_error.iloc[34] = "xx123"

try:
    pd.to_datetime(strings_error, format="%Yxx%mxx%d")
except Exception as e:
    print(e)

parsed = pd.to_datetime(strings_error,
                        format="%Yxx%mxx%d", errors='coerce')
parsed.iloc[34]

parsed = pd.to_datetime(strings_error,
                        format="%Yxx%mxx%d", errors='ignore')
parsed.head()

birth_dates = pd.Series(master.index, index=master.birthDate)
birth_dates.head(2)

birth_dates.index

birth_dates = birth_dates.sort_index()
birth_dates.head(3)

born_1980 = birth_dates['1980']
born_1980.head()

birth_dates['1980-07']

birth_dates['1980-07-24']

birth_dates['1980-07-20':'1980-07-31']

# Other Data Types

diff = birth_dates.index[1] - birth_dates.index[0]
diff

birth_dates.index[10]

birth_dates.index[10] + diff

birth_dates.index - birth_dates.index[0]

diff.components

diff = diff + '5H' - '10M'
diff

diff.round('10D')

diff.round('1H')

birth_dates = birth_dates.to_period(freq="D")
birth_dates.index

birth_dates.index[0]

start = birth_dates.index[0].start_time
start

end = birth_dates.index[0].end_time
end

start < (start + pd.Timedelta('5H')) < end

start < (start + pd.Timedelta('25H')) < end

birth_dates.index.year

birth_dates['1975'].head(2)

birth_dates["1975-01-09"]

# Advanced Strategies

birth_dates.resample('1M')

birth_dates.resample('1M').count().head()

birth_dates.resample('2M').count().head(2)

birth_dates.resample('1Q').count().head(2)

birth_dates.resample('W').count().head(2)

from pandas.tseries.offsets import BDay
p = birth_dates.index[2]
print(p.to_timestamp())
print(p.to_timestamp() + BDay(7))

birth_dates.head()

birth_dates.shift(1, freq="D").head()

birth_dates.shift(1).head()

team_splits.head()

team_splits = team_splits.reset_index()
team_splits.head()

string_date = team_splits.apply(lambda x:
                                pd.Period("{}-{}".format(x.year, x.month)),
                                axis=1)
string_date[0]

team_splits = team_splits.assign(month = string_date)
team_splits = team_splits.drop("year", axis=1)
team_splits.head()

team_splits = team_splits.set_index(["month"])

quarter_sums = team_splits.groupby("name").resample("Q").sum()
quarter_sums.head()

quarter_means = team_splits.groupby("name").resample("Q").mean()
quarter_means.head()

team_splits.to_pickle(os.path.join('..', 'team_splits_periods.pickle'))
birth_dates.to_pickle(os.path.join('..', 'birth_dates.pickle'))
