==> Avanced Pandas/08 Plotting and Presenting Your Data.ipynb <==
import pandas as pd
import os

team_splits = pd.read_pickle(os.path.join('..', 'team_splits_periods.pickle'))
scoring = pd.read_pickle(os.path.join('..', 'scoring.pickle'))

ducks = team_splits.loc[team_splits["name"] == "Anaheim Ducks", :]
ducks = ducks.sort_index()
ducks.head(10)

ducks.resample("A").sum()

ducks.index = ducks.to_timestamp().index.map(
    lambda x: 
    x + pd.DateOffset(years=1) if x.month < 5
    else x
)
ducks = ducks.sort_index()
ducks.head()

resampler = ducks.resample("A-JUN")
resampler

resampler.sum()

# We can't to do this: resampler.expanding().sum()
res = resampler.apply(
        lambda x: x.reset_index()[["W", "L"]]
                   .expanding().sum()
)
res.head(8)

res.index = ducks.index
res.head()

final = res.asfreq(pd.tseries.offsets.MonthBegin())
final.head(10)

fdf = final.reset_index()
fdf.head()

from bokeh.plotting import figure, output_notebook, show
output_notebook()

from bokeh.models.sources import ColumnDataSource as cds

source = cds(fdf)
p = figure(x_axis_type="datetime", height=500, width=500)
p.line(source=source, x='month', y="W", color="green")
p.line(source=source, x='month', y="L", color="red");

show(p)

source = cds(fdf)
p = figure(x_axis_type="datetime", height=500, width=500)
p.line(source=source, x='month', y="W", color="green", legend="Wins")
p.line(source=source, x='month', y="L", color="red", legend="Loses")
p.legend.click_policy="hide";

show(p)

source = cds(fdf)
p = figure(x_axis_type="datetime", height=500, width=500)
p.line(source=source, x='month', y="W", color="green", legend="Wins")
p.line(source=source, x='month', y="L", color="red", legend="Loses")
p.legend.click_policy="hide"
p.title.text = "Anaheim Ducks Performance by Season"
p.title.text_font_size = "25px"
p.title.align = "center";

show(p)

wayne = scoring.set_index("playerID").loc['gretzwa01']
wayne.head()

wayne = wayne[['year', 'tmID', 'GP', 'Pts']]
wayne = wayne.assign(pts_per_game = lambda x: x['Pts'] / x['GP'])
wayne.head()

wayne.tmID.head()

wayne.loc[:, "tmID"] = wayne.tmID.cat.remove_unused_categories()
wayne.tmID.head()

wayne.loc[:, "tmCode"] = wayne["tmID"].cat.codes
wayne.sample(5)

wayne.loc[:, "year"] = pd.to_datetime(wayne.year, format="%Y")

gp_max = wayne.GP.max()
gp_min = wayne.GP.min()
pts_per_game_max = wayne.pts_per_game.max()
pts_per_game_min = wayne.pts_per_game.min()

wayne.loc[:, "height"] = wayne.GP / gp_max
wayne.loc[:, "bottom"] = wayne.tmCode
wayne.loc[:, "top"] = wayne.bottom + wayne.height
wayne.head()

src = cds(wayne)
p = figure(x_axis_type='datetime',
           height=500, width=500,
           y_range=list(wayne.tmID.cat.categories))
p.vbar(bottom='bottom',
       top='top', x='year',
       width=1, source=src);

show(p)

p = figure(x_axis_type='datetime',
           height=500, width=500,
           y_range=list(wayne.tmID.cat.categories))
p.vbar(bottom='bottom',
       top='top', x='year',
       width=1, source=src)
p.xaxis.major_label_text_font_size = "12pt"
p.yaxis.major_label_text_font_size = "12pt"

show(p)

reordered = wayne.sort_values('year').tmID.unique()
reordered

wayne.loc[:, "tmID"] = wayne.tmID.cat.reorder_categories(reordered)
wayne.loc[:, "tmCode"] = wayne["tmID"].cat.codes
wayne.loc[:, "bottom"] = wayne.tmCode
wayne.loc[:, "top"] = wayne.bottom + wayne.height

width = 320 * 24 * 60 * 60 * 1000

src = cds(wayne)
p = figure(x_axis_type='datetime',
           height=500, width=500,
           y_range=list(wayne.tmID.cat.categories))
p.vbar(bottom='bottom',
       top='top', x='year',
       width=width, source=src)
p.xaxis.major_label_text_font_size = "12pt"
p.yaxis.major_label_text_font_size = "12pt"

show(p)

wayne.loc[:, "bottom"] = wayne.tmCode - (wayne.height / 2.0 ) + 0.5
wayne.loc[:, "top"] = wayne.bottom + wayne.height

src = cds(wayne)
p = figure(x_axis_type='datetime',
           height=500, width=500,
           y_range=list(wayne.tmID.cat.categories))
p.vbar(bottom='bottom',
       top='top', x='year',
       width=width, source=src)
p.xaxis.major_label_text_font_size = "12pt"
p.yaxis.major_label_text_font_size = "12pt"

show(p)

wayne.loc[:, "height"] = wayne.GP / gp_max * 0.8
wayne.loc[:, "bottom"] = wayne.tmCode - (wayne.height / 2.0 ) + 0.5
wayne.loc[:, "top"] = wayne.bottom + wayne.height

src = cds(wayne)
p = figure(x_axis_type='datetime',
           height=500, width=500,
           y_range=list(wayne.tmID.cat.categories))
p.vbar(bottom='bottom',
       top='top', x='year',
       width=width, source=src)
p.xaxis.major_label_text_font_size = "12pt"
p.yaxis.major_label_text_font_size = "12pt"

show(p)

from bokeh.models import LinearColorMapper
color_mapper = LinearColorMapper(
    palette="Plasma256", 
    low=pts_per_game_min, 
    high=pts_per_game_max)

src = cds(wayne)
p = figure(x_axis_type='datetime', 
           height=500, width=500,
           y_range=list(wayne.tmID.cat.categories))
p.vbar(bottom='bottom', top='top',
       x='year', width=width,
       color={'field': 'pts_per_game', 'transform': color_mapper},
       source=src);

show(p)

from bokeh.models import ColorBar
src = cds(wayne)
p = figure(x_axis_type='datetime', 
           width=500,
           height=500, 
           y_range=list(wayne.tmID.cat.categories)
)
p.vbar(bottom='bottom', top='top', x='year', width=width,
       color={'field': 'pts_per_game', 'transform': color_mapper}, source=src)
color_bar = ColorBar(color_mapper=color_mapper,
                     label_standoff=8, border_line_color=None, 
                     location=(0,0)
)

p.add_layout(color_bar, 'right');

show(p)

from bokeh.models.tools import HoverTool
hover = HoverTool(
    tooltips=[
        ("Points per game", "@pts_per_game"),
        ("Games played", "@GP")
    ]
)

src = cds(wayne)
p = figure(x_axis_type='datetime', 
           width=500,
           height=500, 
           y_range=list(wayne.tmID.cat.categories)
)
p.vbar(bottom='bottom', top='top', x='year', width=width,
       color={'field': 'pts_per_game', 'transform': color_mapper}, source=src)
color_bar = ColorBar(color_mapper=color_mapper,
                     label_standoff=8, border_line_color=None, 
                     location=(0,0)
)

p.add_layout(color_bar, 'right')
p.add_tools(hover);

show(p)
